+++ packages/plugins/windows/CHANGELOG.md
@@ -2,4 +2,5 @@
 All notable changes to this project will be documented in this file.
 
 ## 0.1.0
+- Fix window tracking and align BrowserWindow with ElectronWindow behavior.
 - Allows for spawning HTML pages as windows while maintaining the same configuration as the main window.